version 2.0 1. price and discount feature 2. location feature 3. stock adjustment feature 4. improvements

+++ src/main/java/com/inventory/manager/application/grn/GRNApplicationService.java
@@ -13,13 +13,20 @@ import com.inventory.manager.application.grn.dto.CreateGoodReceiveNoteRequestDTO
 import com.inventory.manager.application.grn.dto.GetGoodReceiveNoteResponseDTO;
 import com.inventory.manager.application.grn.dto.GoodReceiveNoteLineRequestDTO;
 import com.inventory.manager.application.grn.dto.ListGoodReceiveNoteResponseDTO;
+import com.inventory.manager.application.shared.dto.CalculateTotalRequestDTO;
+import com.inventory.manager.application.shared.dto.CalculateTotalResponseDTO;
+import com.inventory.manager.domain.enums.InvoiceType;
 import com.inventory.manager.domain.item.Item;
 import com.inventory.manager.domain.item.ItemService;
+import com.inventory.manager.domain.location.Location;
+import com.inventory.manager.domain.location.LocationService;
 import com.inventory.manager.domain.stockhistory.grn.GoodReceiveNote;
 import com.inventory.manager.domain.stockhistory.grn.GoodReceiveNoteLine;
 import com.inventory.manager.domain.stockhistory.grn.GoodReceiveNoteService;
 import com.inventory.manager.domain.supplier.Supplier;
 import com.inventory.manager.domain.supplier.SupplierService;
+import com.inventory.manager.domain.util.TotalCalculator;
+import com.inventory.manager.exception.CustomExceptionCodes;
 import com.inventory.manager.exception.NotFoundException;
 
 @Service
@@ -39,22 +46,39 @@ public class GRNApplicationService {
     @Autowired
     private GoodReceiveNoteService goodReceiveNoteService;
 
+    @Autowired
+    private GRNSpecification grnSpec;
+
+    @Autowired
+    private TotalCalculator totalCalculator;
+
+    @Autowired
+    private LocationService locationService;
+
     @Transactional
     public Integer createGRN(CreateGoodReceiveNoteRequestDTO requestDTO) {
         logger.info("[ Create GRN :: " + requestDTO + " ]");
 
         Supplier supplier = supplierService.findSupplierById(requestDTO.getSupplierId());
+
+        if (supplier == null) {
+            throw new NotFoundException(CustomExceptionCodes.SUPPLIER_NOT_FOUND.toString(),
+                    "Supplier not found for the id: " + requestDTO.getSupplierId());
+        }
         List<Integer> itemIds = requestDTO.getGoodReceiveNoteLines().stream()
                 .map(GoodReceiveNoteLineRequestDTO::getItemId).collect(Collectors.toList());
         List<Item> items = itemService.findBySupplierAndItemIds(supplier, itemIds);
 
-        if (items.size() != itemIds.size()) {
-            throw new NotFoundException("Item not found for one of the item ids: " + itemIds);
-        }
+        Location location = locationService.findLocationById(requestDTO.getLocationId());
 
-        GoodReceiveNote goodReceiveNote = grnTransformer.toGoodReceiveNote(requestDTO, supplier);
+        grnSpec.isSatisfiedBy(requestDTO, items, itemIds, location);
+
         List<GoodReceiveNoteLine> goodReceiveNoteLines = grnTransformer.toGoodReceiveNoteLines(
                 requestDTO.getGoodReceiveNoteLines(), items);
+        goodReceiveNoteLines = totalCalculator.updateGRNLineTotal(goodReceiveNoteLines);
+
+        GoodReceiveNote goodReceiveNote = grnTransformer.toGoodReceiveNote(requestDTO, supplier, location);
+        goodReceiveNote = totalCalculator.updateGRNTotal(goodReceiveNoteLines, goodReceiveNote);
 
         return goodReceiveNoteService.createGoodReceiveNote(goodReceiveNote, goodReceiveNoteLines);
     }
@@ -96,7 +120,17 @@ public class GRNApplicationService {
             throw new NotFoundException("GRN not found for the id: " + id);
         }
 
+        if (grn.getLocation().getIsDeleted()) {
+            throw new NotFoundException("Location : '" + grn.getLocation()
+                    + "' was deleted, therefore cannot delete the GRN");
+        }
+
         goodReceiveNoteService.delete(grn);
     }
 
+    public CalculateTotalResponseDTO calculateTotal(CalculateTotalRequestDTO requestDTO) {
+
+        return totalCalculator.calculateTotal(requestDTO, InvoiceType.GRN);
+
+    }
 }
